Implement 'View' drop-down menu.
Allow viewing the map at the same location on external map providers
(Google, Bing, others are easy to add), using a new 'View' drop-down
menu.

Note that the map type (tiled roadmap / satellite) for Google Maps is
not working properly despite the URLs following the official
documentation, this is a known problem:
https://www.reddit.com/r/GoogleMaps/comments/1f9y465/how_to_force_a_map_type_roadsatellite_in_a/

Official documentation is at:
https://developers.google.com/maps/documentation/urls/get-started

Obviously this patch is a first draft and main project maintainers
should feel free to rework this so it suits them.